Task 1 bug fix + Task 2
hello(), and exit() functions has been implemented of Task 2.

A bug detected that if whats stored in the memory as input was not changed, the same command may run again in an empty "Enter" instance afterwards. Therefore a method was introduced to resolve this issue by tricky resetting of whats stored in the memory.